Store ParseErr on heap (#406)
* Box ParseErr

This is because ParseErr is relatively large.
Now it is stored on the heap, and the compiler
also doesn't have to create extra space for the OK
value.

* Remove unnecessary casts to usize

* Remove cast to usize

+++ src/common/position.rs
@@ -73,13 +73,13 @@ impl Position {
     /// the source code.
     ///
     /// Width is always 1 or greater.
-    pub fn get_width(&self) -> i32 {
+    pub fn get_width(&self) -> usize {
         max(
             1,
             max(
                 self.end.pos as i32 - self.start.pos as i32,
                 self.start.pos as i32 - self.end.pos as i32,
-            ),
+            ) as usize,
         )
     }
